Move OpiWidget into separate file
